% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_action_info.R
\docType{class}
\name{ActionInfo-class}
\alias{ActionInfo-class}
\alias{ActionInfo}
\title{A Reference Class to represent Action information}
\description{
A Reference Class to represent Action information
}

