% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_import_file_metadata.R
\name{addRow}
\alias{addRow}
\title{appends row to dataframe from a named list}
\usage{
addRow(dataframe, namedList)
}
\arguments{
\item{namedList}{}
}
\value{
dataframe with additional row with elements assigned to columns by names
}
\description{
appends row to dataframe from a named list
}
