% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_import_file_metadata.R
\name{createFilesMetadataFrame}
\alias{createFilesMetadataFrame}
\title{creates dataframe for explicit metadata for files}
\usage{
createFilesMetadataFrame()
}
\value{
dataframe with columns "access","filePath","fileHash","repoType","repoRoot","repoPath","repoVersion","repoModified","repoModifiedBy","qualityChecked","description"
}
\description{
creates dataframe for explicit metadata for files
}
