% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_get_action_details.R
\name{determineActivityId}
\alias{determineActivityId}
\title{determineActivityId}
\usage{
determineActivityId(repoPath = "", filePath = "")
}
\arguments{
\item{repoPath}{optional: path of folder or file in repository}

\item{filePath}{optional: local path of folder or file in checkout folder}
}
\description{
determineActivityId
}
