% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_callstack_functions.R
\name{getCallInfo}
\alias{getCallInfo}
\title{getCallInfo
returns information about a calling function}
\usage{
getCallInfo(offset = 0)
}
\arguments{
\item{offset}{0 for function directly calling getCallInfo, n for function n levels above in call hierarchy}
}
\value{
list with elements for envName, funcName,
scriptName: contains name of scriptfile, where file of script is given for called function, contains name of package of function otherwise;
scriptPath: contains path of scriptfile or package details respectively (see scriptName)
}
\description{
getCallInfo
returns information about a calling function
}
