% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_callstack_functions.R
\name{getCallStack}
\alias{getCallStack}
\title{get callStack as dataframe}
\usage{
getCallStack(offset = 0, excludeEnvironments = c("base", "methods"))
}
\arguments{
\item{offset}{\itemize{
\item to cut the last rows of the callstack
}}

\item{excludeEnvironments}{("implicit" functions from these environments are not included into the callstack)}
}
\value{
dataframe with rows per call, top level call in row 1; and columns envName, funcName,
scriptName: contains name of scriptfile, where file of script is given for called function, contains name of package of function otherwise;
scriptPath: contains path of scriptfile or package details respectively (see scriptName)
}
\description{
Returns information about the current callstack adjusted for "implicit" function calls like tryCatch.
getCallstack can be used to return information in case of errors or to determine the calling function when creating an actionInfo
}
