% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_callstack_functions.R
\name{getCallStackAsList}
\alias{getCallStackAsList}
\title{getCallStackAsList}
\usage{
getCallStackAsList(offset = 0, excludeEnvironments = c("base",
  "methods"))
}
\arguments{
\item{offset}{\itemize{
\item to cut the last rows of the callstack
}}

\item{excludeEnvironments}{("implicit" functions from these environments are not included into the callstack)}
}
\description{
getCallStackAsList
}
