% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_callstack_functions.R
\name{getCallStackAsString}
\alias{getCallStackAsString}
\title{getCallStackAsString}
\usage{
getCallStackAsString(offset = 0)
}
\arguments{
\item{offset}{\itemize{
\item to cut the last rows of the callstack
}}
}
\value{

}
\description{
getCallStackAsString
}
