% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_callstack_functions.R
\name{getEnvironmentName}
\alias{getEnvironmentName}
\title{getEnvironmentName}
\usage{
getEnvironmentName(env)
}
\arguments{
\item{env}{enviroment}
}
\value{
name of environment
}
\description{
getEnvironmentName
}
