% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_callstack_functions.R
\name{setGlobalStackInfos}
\alias{setGlobalStackInfos}
\title{setGlobalStackInfos}
\usage{
setGlobalStackInfos()
}
\description{
For test purposes only: provides information about R calls, frames, functions where this function is called to the global variable STACKINFOS
}
