% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_system_info.R
\docType{class}
\name{SystemInfo-class}
\alias{SystemInfo-class}
\alias{SystemInfo}
\title{A Reference Class to represent System information}
\description{
A Reference Class to represent System information
}

