% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_metadata_capture - Kopie.R,
%   R/t_metadata_capture.R
\name{tEndMetadataCapture}
\alias{tEndMetadataCapture}
\title{tEndMetadataCapture
Ends metadata capture and saves metadata to File or DB as specified by user}
\usage{
tEndMetadataCapture(storageMode = c("DB", "File"), outputFolder = "./",
  jsonFileName = "", useTimeStamp = T)

tEndMetadataCapture(storageMode = c("DB", "File"), outputFolder = "./",
  jsonFileName = "", useTimeStamp = T)
}
\arguments{
\item{storageMode}{should be one of "None", "File" or "DB"}

\item{outputFolder}{where to store error logfile, json File}

\item{jsonFileName}{name of metadata output file (only for storageMode = "File")}

\item{useTimeStamp}{whether timestamp should be used as filename suffix (only for storageMode = "File")}

\item{storageMode}{should be one of "None", "File" or "DB"}

\item{outputFolder}{where to store error logfile, json File}

\item{jsonFileName}{name of metadata output file (only for storageMode = "File")}

\item{useTimeStamp}{whether timestamp should be used as filename suffix (only for storageMode = "File")}
}
\description{
tEndMetadataCapture
Ends metadata capture and saves metadata to File or DB as specified by user

tEndMetadataCapture
Ends metadata capture and saves metadata to File or DB as specified by user
}
\examples{
tEndMetadataCapture(storageMode = "File", outputFolder="~/my_working_directory/", jsonFileName="my_project",useTimeStamp=T)
tEndMetadataCapture(storageMode = "DB", outputFolder="~/my_working_directory/")
tEndMetadataCapture(storageMode = "File", outputFolder="~/my_working_directory/", jsonFileName="my_project",useTimeStamp=T)
tEndMetadataCapture(storageMode = "DB", outputFolder="~/my_working_directory/")
}
