% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_import_file_metadata - Kopie.R,
%   R/t_import_file_metadata.R
\name{tImportFile}
\alias{tImportFile}
\title{creates action for file import from source to target}
\usage{
tImportFile(importActionDescription = "", sourceRepoType = "Unknown",
  sourceFilePath = "Unknown", sourceFileHash = "Unknown",
  sourceRepoVersion = "Unknown", sourceRepoRoot = "",
  sourceModified = tlconst$OT, sourceModifiedBy = "",
  sourceQualityChecked = F, sourceDescription = "",
  targetRepoType = "Unknown", targetFilePath = "Unknown",
  targetFileHash = "Unknown", targetRepoVersion = "Unknown",
  targetRepoRoot = "", targetModified = tlconst$OT,
  targetModifiedBy = "", targetQualityChecked = F,
  targetDescription = "", offset = 0)

tImportFile(importActionDescription = "", sourceRepoType = "Unknown",
  sourceFilePath = "Unknown", sourceFileHash = "Unknown",
  sourceRepoVersion = "Unknown", sourceRepoRoot = "",
  sourceModified = tlconst$OT, sourceModifiedBy = "",
  sourceQualityChecked = F, sourceDescription = "",
  targetRepoType = "Unknown", targetFilePath = "Unknown",
  targetFileHash = "Unknown", targetRepoVersion = "Unknown",
  targetRepoRoot = "", targetModified = tlconst$OT,
  targetModifiedBy = "", targetQualityChecked = F,
  targetDescription = "", offset = 0)
}
\arguments{
\item{importActionDescription}{optional: any identifying and/or valuable information about the import action}

\item{sourceRepoType}{optional: should contain information about the type of the source repository, if available}

\item{sourceFilePath}{path including the source file name; just the source file name, if no path is available}

\item{sourceFileHash}{md5 file hash of source file}

\item{sourceRepoVersion}{optional: version of source file in source repository, if available}

\item{sourceRepoRoot}{optional: root path of the source repository, if available}

\item{sourceModified}{optional: timestamp of last change of source file, if available}

\item{sourceModifiedBy}{optional: username or name or similar of provider of source file}

\item{sourceQualityChecked}{flag, if the source file is confirmed by the provider to be ready for usage in GxP workflows}

\item{targetRepoType}{information about the type of the target repository}

\item{targetFilePath}{path including the target file name in the repository, e.g. "https://svnserver/svn/reponame/trunk/folder/file.csv"}

\item{targetFileHash}{md5 file hash of target file}

\item{targetRepoVersion}{version of the checked in target file in the target version control system, e.g. svn}

\item{targetRepoRoot}{repository path, e.g. "https://svnserver/svn/reponame"}

\item{targetModified}{check in timestamp in target repository, e.g. available via svninfo}

\item{targetModifiedBy}{user name or name of importing person}

\item{targetQualityChecked}{flag, if the target file is confirmed by the provider or by the importing person to be ready for usage in GxP workflows}

\item{offset}{should be overridden only in special cases}

\item{targetdescription}{}

\item{sourcedescription}{optional: any valuable additional information available for the source file}

\item{targetdescription}{optional: any valuable additional information available for the target file}
}
\description{
besides a description for the action only metadata for the source and the target file and location are passed.
}
