% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_wrapper_functions.R
\name{tReadCsv}
\alias{tReadCsv}
\title{Wrapper function for read.csv to capture metadata}
\usage{
tReadCsv(file, ...)
}
\arguments{
\item{...}{pass on other arguments as you use the read.csv}
}
\value{

}
\description{
Wrapper function for read.csv to capture metadata
}
\examples{
tReadCsv("filepath")
}
