% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_wrapper_functions.R
\name{tReadTable}
\alias{tReadTable}
\title{Wrapper function for read.csv to capture metadata}
\usage{
tReadTable(file, ...)
}
\arguments{
\item{...}{pass on other arguments as you use the read.table}
}
\description{
Wrapper function for read.csv to capture metadata
}
\examples{
tReadTable("filepath")
}
