% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_save_metadata.R
\name{tSaveMetadataToDB}
\alias{tSaveMetadataToDB}
\title{tSaveMetadataToDB
Writes captured  metadata to FlowTrace Database}
\usage{
tSaveMetadataToDB(outputFolder = "./")
}
\arguments{
\item{outputFolder}{path where error file(if generated) should be written}
}
\description{
tSaveMetadataToDB
Writes captured  metadata to FlowTrace Database
}
\examples{
tSaveMetadataToDB(outPutFolder = "./my_project_folder/")
}
