% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_save_metadata.R
\name{tSaveMetadataToFile}
\alias{tSaveMetadataToFile}
\title{tSaveMetadataToFile
Saves captured metadata information to a json file}
\usage{
tSaveMetadataToFile(outputFolder = "./", jsonFileName = "",
  useTimeStamp = T)
}
\arguments{
\item{outputFolder}{directory path where file should be written, defaults to current working directory if not provided}

\item{jsonFileName}{filename}

\item{useTimeStamp}{timestamp as suffix to filename}
}
\value{
json metadata file
}
\description{
tSaveMetadataToFile
Saves captured metadata information to a json file
}
\examples{
tSaveMetadataToFile(outputFolder="~/my_working_directory/", jsonFileName="my_project",useTimeStamp=T)
}
