% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_action_capture.R
\name{tStartAction}
\alias{tStartAction}
\title{Starts meta data capture for an action}
\usage{
tStartAction(
  actionType = "Other",
  actionName = NA,
  description = NA,
  offset = 0
)
}
\arguments{
\item{actionType}{Should be one of DataImport, TLFGeneration, ReportGeneration, Edit, Move, Copy, AnalysisFileGeneration, Analysis or Other}

\item{offset}{optional: if action should be defined by a calling function of higher level}
}
\value{

}
\description{
Starts meta data capture for an action
}
\examples{
tStartAction(actionType = "DataImport")
}
