% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_metadata_capture.R
\name{tStartMetadataCapture}
\alias{tStartMetadataCapture}
\title{tStartMetadataCapture
Starts meta data capture, Initiates global objects for file, action and system infos}
\usage{
tStartMetadataCapture(
  metaDataCapture = T,
  repoPath = "",
  filePath = "",
  actionType = "Run",
  offset = 0
)
}
\arguments{
\item{repoPath}{optional: path of folder or file in repository to determine the PMx activity}

\item{filePath}{optional: local path of folder or file in checkout folder to determine the PMx activity}

\item{actionType}{should be one of Run, DataImport, TLFGeneration, ReportGeneration, Edit, Move, Copy, AnalysisFileGeneration, Analysis or Other}

\item{offset}{optional: if action should be defined by a calling function of higher level}
}
\value{

}
\description{
tStartMetadataCapture
Starts meta data capture, Initiates global objects for file, action and system infos
}
\examples{
tStartMetadataCapture(metaDataCapture = T)
}
