% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_import_file_metadata.R
\name{tStoreActionMetadata}
\alias{tStoreActionMetadata}
\title{creates explicit metadata for action}
\usage{
tStoreActionMetadata(
  actionType = "Other",
  actionDescription = "",
  actionName = "",
  executedBy = "",
  executed = tlconst$OT,
  executionStarted = tlconst$OT,
  inputFilePaths,
  outputFilePaths,
  filesMetadataFrame,
  offset = 0
)
}
\arguments{
\item{actionType}{should be one of DataImport, TLFGeneration, ReportGeneration, Edit, Move, Copy, AnalysisFileGeneration, Analysis or Other}

\item{actionDescription}{optional}

\item{actionName}{optional: should be passed only in special cases, is normally captured automatically}

\item{executedBy}{optional: should be passed only in special cases, is normally captured automatically}

\item{executed}{optional: should be passed only in special cases, is normally captured automatically}

\item{executionStarted}{optional}

\item{inputFilePaths}{vector of input filepaths; only one of inputFilePaths and inputFilesMetadataFrame may be given}

\item{outputFilePaths}{vector of output filepaths; only one of outputFilePaths and filesMetadataFrame may be given}

\item{filesMetadataFrame}{dataframe of output files metadata (see: createFilesMetadataFrame);only one of inputFilePaths/outputFilePaths and filesMetadataFrame may be given}

\item{offset}{}
}
\value{

}
\description{
creates explicit metadata for action
}
