% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_import_file_metadata.R
\name{tStoreFileMetadata}
\alias{tStoreFileMetadata}
\title{Stores file metadata}
\usage{
tStoreFileMetadata(access = c("read", "write"), filePath = NA,
  fileHash = NA, repoType = NA, repoRoot = NA, repoPath = NA,
  repoVersion = NA, repoModified = NA, repoModifiedBy = NA,
  qualityChecked = NA, description = NA)
}
\arguments{
\item{access}{Should be one of read or Write}

\item{filePath}{local file path; if this is not given, the following parameters except for repoRoot, qualityChecked, description are mandatory}

\item{fileHash}{optional: only used, if filePath is not given}

\item{repoType}{optional:}

\item{repoRoot}{optional:}

\item{repoPath}{optional:}

\item{repoVersion}{optional:}

\item{repoModified}{optional:}

\item{repoModifiedBy}{optional:}

\item{qualityChecked}{optional:}

\item{description}{optional:}
}
\description{
this function can be called explicitly from users of traceLib if files are read or written by functions, where no t_wrapper is available
}
\examples{
tStoreFileMetadata(access = "read", filePath = "filepath")
}
