% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_wrapper_functions.R
\name{tWriteCsv}
\alias{tWriteCsv}
\title{Wrapper function for write.csv to capture metadata}
\usage{
tWriteCsv(x, file, ...)
}
\arguments{
\item{x}{Data object}

\item{file}{filepath}

\item{...}{other arguments to write.csv}
}
\description{
Wrapper function for write.csv to capture metadata
}
\examples{
tWriteCsv(x = data, file = "filepath")
}
