% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_read_write_json.R
\name{writeInfosToJson}
\alias{writeInfosToJson}
\title{Title}
\usage{
writeInfosToJson(fileInfos, actionInfos, systemInfo, outputFolder,
  jsonFileName, useTimeStamp)
}
\arguments{
\item{outputFolder}{}
}
\description{
Title
}
